Added ldlasso class validities and ldlassoSolution class definition
Here I defined the ldlassoSolution class as an class inherited from ldlasso which contains a slot for the solution, beta.  I also defined some validities for the ldlasso class, namely: s1, s2 need to be greater than zero, delta has to be between 0 and 1e-6, all entries in geno must be 0, 1, or 2, and all entries in pheno must be 0 or 1.