Added ldlassoModel.R
I also cleaned up the validity so that there is no error checking in ldlassoSolve.  I think it was causing problems and it makes more sense in the validity file anyway.  Note that I also renamed functions.R to ldlassoSolve.R.  For the most part I want to have one file per function.  It makes searching for functions easier, and I think is a more straight-forward layout.