Created findS1 method and generic.
I had trouble getting this to work, and it seems that the missing piece was the generic function.  The conclusion is that there must be a generic function associated with the method.  I should check this assumption out.  The findS1 method returns the whole ldlasso object with the new s1 value and the solution, bete, nulled out.