Changed so that we re-define alpha if 2/number of SNPs is greater than alpha.
When the number of SNPs is low, say around 24, the bisection algorithm nac get caught in a loop if 1/number of SNPs is not sufficiently close to alpha.  Since at least one SNP is always included in final model we needed to decrease alpha to be exactly 1/alpha to avoid the inifinte loop.