Fixed findS1 method to reinstanciate ldlasso objects so that OR is updated.
It is not enough to only update slots.  Becuase we now compute auxilliary stats like OR, and maf we need to recreate ldlasso objects every time a slot such as geno or pheno is changed.  We also included the validity test at the end.  It is okay to violate the LD LASSO constraint during the search for s1, but we cannot return an invalid solution.  The LD LASSO constraint violation is a result of the faulty solve.QP.  The error is reflected in the err value that is output to log file.